requirements.txt
alembic==1.8.1
amqp==2.6.1
aniso8601==9.0.1
arrow==0.10.0
attrs==22.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
Flask==2.2.2
Flask-Cors==3.0.3
Flask-JWT-Extended==4.2.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.9
Flask-Script==1.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==4.6.11
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
minio==7.1.3
more-itertools==9.0.0
openpyxl==3.0.10
packaging==21.3
pluggy==0.6.0
py==1.11.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==3.5.1
pytest-cov==2.5.1
pytest-flask==0.10.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
pytz==2022.6
redis==3.0.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
uWSGI==2.0.20
vine==1.3.0
Werkzeug==2.2.2
mysql-connector-python==8.0.31
openai==0.27.6
